+++ zoo/src/test/scala/com/intel/analytics/zoo/pipeline/api/net/TFNetSpec.scala
@@ -137,14 +137,4 @@ class TFNetSpec extends FlatSpec with Matchers with BeforeAndAfter {
     gradInput.size() should be (input.size())
   }
 
-  "TFNet" should "load keras" in {
-    val inputs = "reshape_input:0"
-    val outputs = "output/Softmax:0"
-
-    val model = TFNet("/tmp/mytf/frozen_inference_graph.pb", Array(inputs), Array(outputs))
-    model.forward(Tensor[Float].range(0, 676 * 2 - 1).resize(2, 676).div(1000))
-
-    println(model.output)
-  }
-
 }
